--- vaadin-combo-box-flow-parent/vaadin-combo-box-flow/src/test/java/com/vaadin/flow/component/combobox/HasLabelTest.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2000-2022 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.component.combobox;
-
-import com.vaadin.flow.component.HasLabel;
-import org.junit.Assert;
-import org.junit.Test;
-
-public class HasLabelTest {
-
-    @Test
-    public void comboBox() {
-        ComboBox<String> c = new ComboBox<>();
-        Assert.assertTrue(c instanceof HasLabel);
-    }
-
-}
